% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_donations.R
\name{get_donations}
\alias{get_donations}
\title{Get donations}
\usage{
get_donations()
}
\value{
A data.frame. Variable names are cleaned.
}
\description{
This function gets live data about donations from the Electoral Commission donations database.
}
\examples{
get_donations()
}
