% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loans.R
\name{get_loans}
\alias{get_loans}
\title{Get loans}
\usage{
get_loans()
}
\value{
A data.frame. Variable names are cleaned.
}
\description{
This function gets live data about loans from the Electoral Commission donations database.
}
\examples{
get_loans()
}
