% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spending.R
\name{get_spending}
\alias{get_spending}
\title{Get spending}
\usage{
get_spending()
}
\value{
A data.frame. Variable names are cleaned.
}
\description{
This function gets live data about spending from the Electoral Commission spending database.
}
\examples{
get_spending()
}
