% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_donations.R
\name{get_donations}
\alias{get_donations}
\title{Get donations}
\usage{
get_donations()
}
\value{
A data.frame.
}
\description{
This function gets donations from the electoral commission website.
}
\examples{
get_donations()
}
